+++ Gui/AchievementTooltip/NotObtainable.lua
@@ -1,5 +1,6 @@
 local _, addon = ...;
 local section = {};
+tinsert(addon.Gui.AchievementTooltip.Sections, section);
 
 function section.CheckAdd(achievement)
     return achievement.TemporaryObtainable;
@@ -9,5 +10,3 @@ function section.Add(achievement)
 	local text, color = addon.Data.TemporaryObtainable:GetNotObtainableText(achievement);
 	GameTooltip:AddLine(text, color.R, color.G, color.B);
 end
-
-tinsert(addon.Gui.AchievementTooltip.Sections, section);